chore: remove old vote and login logic
